SYS-641 fix mythtv-backend config.xml template substitution
Restore template variable substitution for config.xml in the mythtv-backend
entrypoint script. This was accidentally removed in commit 451463f when
transitioning to MythTV 35.0 and removing Apache/MythWeb support.

The config.xml template contains placeholder variables that must be replaced
with actual environment variable values:
- {{ DBSERVER }} -> $DBSERVER
- {{ DBNAME }} -> $DBNAME
- {{ DBPASSWORD }} -> $DBPASSWORD
- {{ LOCALHOSTNAME }} -> $LOCALHOSTNAME

Without this substitution, the application fails to connect to the database
with errors about connecting to a literal '{{ DBSERVER }}' hostname.

Fixes: Environment variable substitution broken in mythtv-backend 35.0